requirements.txt
yt-dlp
groq
chromadb
sentence-transformers
fastapi>=0.115.2
uvicorn>=0.14.0
sqlalchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.13.1
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
cors>=1.0.1
httpx>=0.25.2
audioop-lts